Aggiunti file room * collegare linegraph a room per database -5 min * collegare maincativity a room * rendere più leggibili scritte in RGB Values * creare string per italiano
# migliorie
* fotocamera, si può migliorare,
   - controllare `.let` se hai implementato correttamente come su Android Developers
* correggere status bar (modalità scura/chiara-land/port)
* controlla progetto github per ulteriori migliorie

+++ app/src/main/java/com/example/hueharvester/ColorData.kt
@@ -5,8 +5,8 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "color_data")
 data class ColorData(
-    @PrimaryKey(autoGenerate = true) val id: Long = 0,
-    val timestamp: Long,
+    @PrimaryKey(autoGenerate = true) val id: Int = 0,
+    val timestamp: Float,
     val red: Int,
     val green: Int,
     val blue: Int
